requirements.txt
boto3==1.24.31
numpy==1.21.5
opencv_python==4.5.4.58
pandas==1.4.2
Pillow==9.2.0
pyparsing==3.0.4
tensorflow==2.3.0
